added players and users table seeding

+++ app/routes.php
@@ -11,73 +11,73 @@
 |
 */
 /* Registration */
-Route::get('register', array('as' => 'auth.register', 'uses' => 'Odeva\Masterpoint\Controller\Auth@getRegister'));
-Route::post('register', 'Odeva\Masterpoint\Controller\Auth@postRegister');
-Route::get('register/success', array('as' => 'auth.register.success', 'uses' => 'Odeva\Masterpoint\Controller\Auth@getRegisterSuccess'));
-Route::get('register/activate/{code}', array('as' => 'auth.register.activate', 'uses' => 'Odeva\Masterpoint\Controller\Auth@getActivate'))
+Route::get('register', array('as' => 'auth.register', 'uses' => 'Odeva\Masterpoint\Controller\Site\Auth@getRegister'));
+Route::post('register', 'Odeva\Masterpoint\Controller\Site\Auth@postRegister');
+Route::get('register/success', array('as' => 'auth.register.success', 'uses' => 'Odeva\Masterpoint\Controller\Site\Auth@getRegisterSuccess'));
+Route::get('register/activate/{code}', array('as' => 'auth.register.activate', 'uses' => 'Odeva\Masterpoint\Controller\Site\Auth@getActivate'))
 	->where('code', '[A-Za-z0-9]+');
 /* Login */
-Route::get('login', array('as' => 'auth.login', 'uses' => 'Odeva\Masterpoint\Controller\Auth@getLogin'));
-Route::post('login', 'Odeva\Masterpoint\Controller\Auth@postLogin');
+Route::get('login', array('as' => 'auth.login', 'uses' => 'Odeva\Masterpoint\Controller\Site\Auth@getLogin'));
+Route::post('login', 'Odeva\Masterpoint\Controller\Site\Auth@postLogin');
 /* Logout */
-Route::get('logout', array('as' => 'auth.logout', 'uses' => 'Odeva\Masterpoint\Controller\Auth@getLogout'));
+Route::get('logout', array('as' => 'auth.logout', 'uses' => 'Odeva\Masterpoint\Controller\Site\Auth@getLogout'));
 /* Reset Password */
-Route::get('reset-password', array('as' => 'auth.login.reset-password', 'uses' => 'Odeva\Masterpoint\Controller\Auth@getResetPassword'));
-Route::post('reset-password', 'Odeva\Masterpoint\Controller\Auth@postResetPassword');
-Route::get('reset-password/success', array('as' => 'auth.login.reset-password.success', 'uses' => 'Odeva\Masterpoint\Controller\Auth@getResetPasswordSuccess'));
-Route::get('reset-password/change/{code}', array('as' => 'auth.login.reset-password.change', 'uses' => 'Odeva\Masterpoint\Controller\Auth@getResetPasswordChange'))
+Route::get('reset-password', array('as' => 'auth.login.reset-password', 'uses' => 'Odeva\Masterpoint\Controller\Site\Auth@getResetPassword'));
+Route::post('reset-password', 'Odeva\Masterpoint\Controller\Site\Auth@postResetPassword');
+Route::get('reset-password/success', array('as' => 'auth.login.reset-password.success', 'uses' => 'Odeva\Masterpoint\Controller\Site\Auth@getResetPasswordSuccess'));
+Route::get('reset-password/change/{code}', array('as' => 'auth.login.reset-password.change', 'uses' => 'Odeva\Masterpoint\Controller\Site\Auth@getResetPasswordChange'))
 	->where('code', '[A-Za-z0-9]+');
-Route::post('reset-password/change/{code}', 'Odeva\Masterpoint\Controller\Auth@postResetPasswordChange')
+Route::post('reset-password/change/{code}', 'Odeva\Masterpoint\Controller\Site\Auth@postResetPasswordChange')
 	->where('code', '[A-Za-z0-9]+');
 
 /* User */
 Route::group(array('prefix' => 'panel'), function() {
-	Route::get('/', array('as' => 'panel', 'uses' => 'Odeva\Masterpoint\Controller\Panel@getIndex'));
-	Route::post('perpage', array('as' => 'panel.perpage', 'uses' => 'Odeva\Masterpoint\Controller\Panel@postPerPage'));
+	Route::get('/', array('as' => 'panel', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Panel@getIndex'));
+	Route::post('perpage', array('as' => 'panel.perpage', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Panel@postPerPage'));
 	/* Account */
 	Route::group(array('prefix' => 'account'), function() {
-		Route::get('/', array('as' => 'panel.account', 'uses' => 'Odeva\Masterpoint\Controller\Account@getIndex'));
-		Route::get('notifications', array('as' => 'panel.account.notifications', 'uses' => 'Odeva\Masterpoint\Controller\Account@getNotifications'));
-		Route::get('details', array('as' => 'panel.account.details', 'uses' => 'Odeva\Masterpoint\Controller\Account@getDetails'));
-		Route::post('details', 'Odeva\Masterpoint\Controller\Account@postDetails');
+		Route::get('/', array('as' => 'panel.account', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Account@getIndex'));
+		Route::get('notifications', array('as' => 'panel.account.notifications', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Account@getNotifications'));
+		Route::get('details', array('as' => 'panel.account.details', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Account@getDetails'));
+		Route::post('details', 'Odeva\Masterpoint\Controller\Panel\Account@postDetails');
 		/* Account Password */
 		Route::group(array('prefix' => 'password'), function() {
-			Route::get('/', array('as' => 'panel.account.password', 'uses' => 'Odeva\Masterpoint\Controller\Account@getPassword'));
-			Route::post('/', 'Odeva\Masterpoint\Controller\Account@postPassword');
-			Route::get('success', array('as' => 'panel.account.password.success', 'uses' => 'Odeva\Masterpoint\Controller\Account@getPasswordSuccess'));
+			Route::get('/', array('as' => 'panel.account.password', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Account@getPassword'));
+			Route::post('/', 'Odeva\Masterpoint\Controller\Panel\Account@postPassword');
+			Route::get('success', array('as' => 'panel.account.password.success', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Account@getPasswordSuccess'));
 		});
 	});
 	/* Club User */
 	Route::group(array('prefix' => 'club'), function() {
-		Route::get('/', array('as' => 'panel.club', 'uses' => 'Odeva\Masterpoint\Controller\Club@getIndex'));
+		Route::get('/', array('as' => 'panel.club', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Club\Club@getIndex'));
 		/* Tournament */
 		Route::group(array('prefix' => 'tournaments'), function() {
-			Route::get('/', array('as' => 'panel.club.tournaments', 'uses' => 'Odeva\Masterpoint\Controller\Tournament@getIndex'));
+			Route::get('/', array('as' => 'panel.club.tournaments', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Club\Tournament@getIndex'));
 		});
 		
 	});
 	/* Admin User */
 	Route::group(array('prefix' => 'admin'), function() {
-		Route::get('/', array('as' => 'panel.admin', 'uses' => 'Odeva\Masterpoint\Controller\Admin@getIndex'));
+		Route::get('/', array('as' => 'panel.admin', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Admin@getIndex'));
 		/* Clubs */
 		Route::model('club', 'Odeva\Masterpoint\Model\Club');
 		Route::group(array('prefix' => 'clubs'), function() {
-			Route::get('/', array('as' => 'panel.admin.clubs.index', 'uses' => 'Odeva\Masterpoint\Controller\Clubs@index'));
-			Route::post('/', array('as' => 'panel.admin.clubs.store', 'uses' => 'Odeva\Masterpoint\Controller\Clubs@store'));
-			Route::get('create', array('as' => 'panel.admin.clubs.create', 'uses' => 'Odeva\Masterpoint\Controller\Clubs@create'));
-			Route::get('{club}/edit', array('as' => 'panel.admin.clubs.edit', 'uses' => 'Odeva\Masterpoint\Controller\Clubs@edit'));
-			Route::put('{club}', array('as' => 'panel.admin.clubs.update', 'uses' => 'Odeva\Masterpoint\Controller\Clubs@update'));
-			Route::delete('{club}', array('as' => 'panel.admin.clubs.destroy', 'uses' => 'Odeva\Masterpoint\Controller\Clubs@destroy'));
+			Route::get('/', array('as' => 'panel.admin.clubs.index', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Clubs@index'));
+			Route::post('/', array('as' => 'panel.admin.clubs.store', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Clubs@store'));
+			Route::get('create', array('as' => 'panel.admin.clubs.create', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Clubs@create'));
+			Route::get('{club}/edit', array('as' => 'panel.admin.clubs.edit', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Clubs@edit'));
+			Route::put('{club}', array('as' => 'panel.admin.clubs.update', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Clubs@update'));
+			Route::delete('{club}', array('as' => 'panel.admin.clubs.destroy', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Clubs@destroy'));
 		});
 		/* Users */
 		Route::model('user', 'Odeva\Masterpoint\Model\User');
 		Route::group(array('prefix' => 'users'), function() {
-			Route::get('/', array('as' => 'panel.admin.users.index', 'uses' => 'Odeva\Masterpoint\Controller\Users@index'));
-			Route::post('/', array('as' => 'panel.admin.users.store', 'uses' => 'Odeva\Masterpoint\Controller\Users@store'));
-			Route::get('create', array('as' => 'panel.admin.users.create', 'uses' => 'Odeva\Masterpoint\Controller\Users@create'));
-			Route::get('{user}/edit', array('as' => 'panel.admin.users.edit', 'uses' => 'Odeva\Masterpoint\Controller\Users@edit'));
-			Route::put('{user}', array('as' => 'panel.admin.users.update', 'uses' => 'Odeva\Masterpoint\Controller\Users@update'));
-			Route::delete('{user}', array('as' => 'panel.admin.users.destroy', 'uses' => 'Odeva\Masterpoint\Controller\Users@destroy'));
+			Route::get('/', array('as' => 'panel.admin.users.index', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Users@index'));
+			Route::post('/', array('as' => 'panel.admin.users.store', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Users@store'));
+			Route::get('create', array('as' => 'panel.admin.users.create', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Users@create'));
+			Route::get('{user}/edit', array('as' => 'panel.admin.users.edit', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Users@edit'));
+			Route::put('{user}', array('as' => 'panel.admin.users.update', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Users@update'));
+			Route::delete('{user}', array('as' => 'panel.admin.users.destroy', 'uses' => 'Odeva\Masterpoint\Controller\Panel\Admin\Users@destroy'));
 		});
 	});
 });
@@ -115,16 +115,16 @@ Route::get('lisans-yetkilisi/vize', 'LicenceUser@getVize');
 
 /* Superadmin */
 /* Site */
-Route::get('/', array('as' => 'home', 'uses' => 'Odeva\Masterpoint\Controller\Home@getIndex'));
-Route::get('menu1', array('as' => 'menu1', 'uses' => 'Odeva\Masterpoint\Controller\Home@getMenu1'));
-Route::get('menu2', array('as' => 'menu2', 'uses' => 'Odeva\Masterpoint\Controller\Home@getMenu2'));
-Route::get('menu2a', array('as' => 'menu2a', 'uses' => 'Odeva\Masterpoint\Controller\Home@getMenu2a'));
-Route::get('menu2b', array('as' => 'menu2b', 'uses' => 'Odeva\Masterpoint\Controller\Home@getMenu2b'));
-Route::get('menu2c', array('as' => 'menu2c', 'uses' => 'Odeva\Masterpoint\Controller\Home@getMenu2c'));
-Route::get('menu3', array('as' => 'menu3', 'uses' => 'Odeva\Masterpoint\Controller\Home@getMenu3'));
-Route::get('menu3a', array('as' => 'menu3a', 'uses' => 'Odeva\Masterpoint\Controller\Home@getMenu3a'));
-Route::get('menu3b', array('as' => 'menu3b', 'uses' => 'Odeva\Masterpoint\Controller\Home@getMenu3b'));
-Route::get('menu4', array('as' => 'menu4', 'uses' => 'Odeva\Masterpoint\Controller\Home@getMenu4'));
+Route::get('/', array('as' => 'home', 'uses' => 'Odeva\Masterpoint\Controller\Site\Home@getIndex'));
+Route::get('menu1', array('as' => 'menu1', 'uses' => 'Odeva\Masterpoint\Controller\Site\Home@getMenu1'));
+Route::get('menu2', array('as' => 'menu2', 'uses' => 'Odeva\Masterpoint\Controller\Site\Home@getMenu2'));
+Route::get('menu2a', array('as' => 'menu2a', 'uses' => 'Odeva\Masterpoint\Controller\Site\Home@getMenu2a'));
+Route::get('menu2b', array('as' => 'menu2b', 'uses' => 'Odeva\Masterpoint\Controller\Site\Home@getMenu2b'));
+Route::get('menu2c', array('as' => 'menu2c', 'uses' => 'Odeva\Masterpoint\Controller\Site\Home@getMenu2c'));
+Route::get('menu3', array('as' => 'menu3', 'uses' => 'Odeva\Masterpoint\Controller\Site\Home@getMenu3'));
+Route::get('menu3a', array('as' => 'menu3a', 'uses' => 'Odeva\Masterpoint\Controller\Site\Home@getMenu3a'));
+Route::get('menu3b', array('as' => 'menu3b', 'uses' => 'Odeva\Masterpoint\Controller\Site\Home@getMenu3b'));
+Route::get('menu4', array('as' => 'menu4', 'uses' => 'Odeva\Masterpoint\Controller\Site\Home@getMenu4'));
 
 //Route::controller('/', 'Odeva\Masterpoint\Controller\HomeController');
 /*
@@ -150,4 +150,4 @@ Route::get('sikca-sorulan-sorular', 'Home@getSikcaSorulanSorular');
  * 
  */
 
-Route::controller('', 'Odeva\Masterpoint\Controller\Home');
+Route::controller('', 'Odeva\Masterpoint\Controller\Site\Home');
